requirements.txt
appdirs==1.4.4
attrs==22.1.0
black==19.3b0
Click==7.0
cycler==0.11.0
kiwisolver==1.4.4
matplotlib==3.6.2
nose==1.3.7
numpy==1.23.5
opencv-contrib-python==4.6.0.66
Pillow==8.2.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
toml==0.10.0
virtualenv==16.7.4
xlrd==1.2.0
xmltodict==0.12.0
yapf==0.28.0
mysql-connector-python==8.0.29
protobuf==4.21.9